chore: try biome formatter
Adds biome as a dev dependency, some scripts, and a config file
which roughly matches the prettier config file.

There are tons of differences due to multi-line ternary indentation
and max line length handling (biome eagerly expands lines to reach
it, I think), but it runs in under 1s on the whole codebase which
is nice.